store: split out assertion support tests to store_asserts_test.go
